backend/requirements.txt
Faker ~= 0.7
Flask ~= 0.12
Flask-Cors ~= 3.0
Flask-SQLAlchemy ~= 1.0
Pillow ~= 4.3
SQLAlchemy ~= 1.1
gunicorn ~= 19.7
psycopg2 ~= 2.7
